Move Serialize/Deserialize impls to separate file
Having them in rope.rs made that file harder to read.
